Add createActivity test in backend
A new backend test file called createActivity.test.ts has been created. This includes tests for the creation of new users and activities, as well as the necessary setup and teardown steps for the database before and after testing. This will help in verifying that the 'activity' creation feature is working as expected.