Update activities model definitions
This commit refines the activities model by adding enumeration for 'type' and implementing some default values for certain fields. It also expands the swagger documentation by providing more precise field descriptions and an illustrative example. The import statements are updated to support these changes. In addition, the verb tense in the swagger descriptions was revised as it narrates past happenings and not future predictions.