Skip activity creation test
The commit marks the 'activity creation' test suite to be skipped within the test execution. The test suite was previously run with every execution but with this change, it will be ignored. This is usually done for non-critical tests or tests that are currently causing issues but will be fixed later.