Add GPX file processing in activities controller
A new method has been added to the activities controller that allows the processing of GPX files. This new function, createActivityGPX, validates the input data, reads the GPX file if it exists, parses it, and then inserts it into the database. It also includes error handling to ensure that no wrong input data is sent to the database.