Refactor GPX file handling in activitiesController
This commit refactors the GPX file handling in activitiesController. It changes the way of importing from GPXParser and moves the details of the GPX to JSON conversion process to the GPXParser, making the controller cleaner. Furthermore, it now passes 'date', 'durationTotal', and 'distanceTotal' from the converted GPX metadata to the database.